backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
Pillow==10.4.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.9.2
aiofiles==24.1.0
httpx==0.27.2
